Passed CantWriteWordThatAlreadyExists and CanDeleteFile tests.

+++ GuessingGameTests/UnitTest1.cs
@@ -9,10 +9,10 @@ namespace GuessingGameTests
         [Fact]
         public void CanReadLines()
         {
+            if (File.Exists(GuessingGame.Program.filepath)) File.Delete(GuessingGame.Program.filepath);
             string[] words = new string[] { "Pink", "Flamingo" };
             File.WriteAllLines(GuessingGame.Program.filepath, words);
             string[] filewords = GuessingGame.Program.ReadWords();
-            File.Delete(GuessingGame.Program.filepath);
             Assert.Equal("Pink", filewords[0]);
         }
 
@@ -31,6 +31,26 @@ namespace GuessingGameTests
             GuessingGame.Program.WriteWord("contemporaneous");
             string[] filewords = File.ReadAllLines(GuessingGame.Program.filepath);
             Assert.Equal("contemporaneous", filewords[0]);
+        }
+
+        [Fact]
+        public void CantWriteWordThatAlreadyExists()
+        {
+            if (File.Exists(GuessingGame.Program.filepath)) File.Delete(GuessingGame.Program.filepath);
+            GuessingGame.Program.WriteWord("bacon");
+            GuessingGame.Program.WriteWord("bacon");
+            string[] filewords = File.ReadAllLines(GuessingGame.Program.filepath);
+            Assert.Single(filewords);
+        }
+
+        [Fact]
+        public void CanDeleteFile()
+        {
+            if (File.Exists(GuessingGame.Program.filepath)) File.Delete(GuessingGame.Program.filepath);
+            string[] words = new string[] { "Pink", "Flamingo" };
+            File.WriteAllLines(GuessingGame.Program.filepath, words);
+            GuessingGame.Program.DeleteWords();
+            Assert.False(File.Exists(GuessingGame.Program.filepath));
         }
     }
 }
